move the hooks file to components
