--- pages/hooks/useFetcher.ts
@@ -1,7 +1,0 @@
-import {  useCallback } from "react";
-import { fetchData } from "../api/fetchData";
-
-export const useFetcher = () => {
-    const data = useCallback(fetchData, [])
-    return {data}
-}
